pyproject.toml
[project]
name = "actividad-de-evaluacion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ipykernel>=6.29.5",
    "azureml-core>=1.60.0",
    "pandas>=2.2.3",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "joblib>=1.4.2",
    "setuptools>=79.0.0",
    "xgboost>=2.1.4",
]
